+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/tests/latest/test_webapp_commands.py
@@ -178,9 +178,9 @@ class WebappQuickCreateTest(ScenarioTest):
         time.sleep(45)  # 45 seconds should be enough for the deployment finished(Skipped under playback mode)
         r = requests.get('http://{}.azurewebsites.net'.format(webapp_name), timeout=240)
         # verify the web page
-        if 'Hello world' not in str(r.content):
+        if 'Your App Service app is up and running' not in str(r.content):
             # dump out more info for diagnose
-            self.fail("'Hello world' is not found in the web page. We get instead:" + str(r.content))
+            self.fail("'Your App Service app is up and running' is not found in the web page. We get instead:" + str(r.content))
 
     @ResourceGroupPreparer(parameter_name='resource_group', parameter_name_for_location='resource_group_location')
     @ResourceGroupPreparer(parameter_name='resource_group2', parameter_name_for_location='resource_group_location2')
@@ -696,6 +696,23 @@ class FunctionappACRDeploymentScenarioTest(ScenarioTest):
         self.cmd('functionapp delete -g {} -n {}'.format(resource_group, functionapp))
 
 
+class FunctionAppReservedInstanceTest(ScenarioTest):
+    @ResourceGroupPreparer()
+    @StorageAccountPreparer()
+    def test_functionapp_reserved_instance(self, resource_group, storage_account):
+        functionapp_name = self.create_random_name('functionappwithreservedinstance', 40)
+        self.cmd('functionapp create -g {} -n {} -c westus -s {} --os-type Windows'
+                 .format(resource_group, functionapp_name, storage_account)).assert_with_checks([
+                     JMESPathCheck('state', 'Running'),
+                     JMESPathCheck('name', functionapp_name),
+                     JMESPathCheck('kind', 'functionapp'),
+                     JMESPathCheck('hostNames[0]', functionapp_name + '.azurewebsites.net')])
+        self.cmd('functionapp config set -g {} -n {} --prewarmed-instance-count 4'
+                 .format(resource_group, functionapp_name)).assert_with_checks([
+                     JMESPathCheck('reservedInstanceCount', 4)])
+        self.cmd('functionapp delete -g {} -n {}'.format(resource_group, functionapp_name))
+
+
 class WebappGitScenarioTest(ScenarioTest):
     @ResourceGroupPreparer()
     def test_webapp_git(self, resource_group):
@@ -935,6 +952,45 @@ class FunctionAppWithPlanE2ETest(ScenarioTest):
         self.cmd('functionapp delete -g {} -n {}'.format(resource_group, functionapp_name))
 
 
+class FunctionUpdatePlan(ScenarioTest):
+    @ResourceGroupPreparer(location='centralus')
+    @StorageAccountPreparer()
+    def test_move_plan_to_elastic(self, resource_group, storage_account):
+        functionapp_name = self.create_random_name('functionappelastic', 40)
+        ep_plan_name = self.create_random_name('somerandomplan', 40)
+        second_plan_name = self.create_random_name('secondplan', 40)
+        s1_plan_name = self.create_random_name('ab1planname', 40)
+
+        plan_result = self.cmd('functionapp plan create -g {} -n {} --sku EP1'.format(resource_group, ep_plan_name), checks=[
+            JMESPathCheck('sku.name', 'EP1')
+        ]).get_output_in_json()
+
+        self.cmd('functionapp plan create -g {} -n {} --sku EP1'.format(resource_group, second_plan_name), checks=[
+            JMESPathCheck('sku.name', 'EP1')
+        ]).get_output_in_json()
+
+        self.cmd('functionapp plan create -g {} -n {} --sku S1'.format(resource_group, s1_plan_name), checks=[
+            JMESPathCheck('sku.name', 'S1')
+        ])
+
+        self.cmd('functionapp create -g {} -n {} --plan {} -s {}'
+                 .format(resource_group, functionapp_name, second_plan_name, storage_account)).assert_with_checks([
+                     JMESPathCheck('state', 'Running'),
+                     JMESPathCheck('name', functionapp_name),
+                     JMESPathCheck('hostNames[0]', functionapp_name + '.azurewebsites.net')])
+
+        self.cmd('functionapp update -g {} -n {} --plan {}'
+                 .format(resource_group, functionapp_name, ep_plan_name)).assert_with_checks([
+                     JMESPathCheck('state', 'Running'),
+                     JMESPathCheck('name', functionapp_name),
+                     JMESPathCheck('serverFarmId', plan_result['id']),
+                     JMESPathCheck('hostNames[0]', functionapp_name + '.azurewebsites.net')])
+
+        # Moving to and from an App Service plan (not Elastic Premium) is not allowed right now
+        self.cmd('functionapp update -g {} -n {} --plan {}'
+                 .format(resource_group, functionapp_name, s1_plan_name), expect_failure=True)
+
+
 class FunctionAppWithConsumptionPlanE2ETest(ScenarioTest):
     @ResourceGroupPreparer(name_prefix='azurecli-functionapp-c-e2e', location='westus')
     @StorageAccountPreparer()
@@ -1072,6 +1128,36 @@ class FunctionAppServicePlan(ScenarioTest):
             JMESPathCheck('sku.name', 'S1')
         ])
 
+    @ResourceGroupPreparer(location='centralus')
+    def test_functionapp_elastic_plan(self, resource_group):
+        plan = self.create_random_name(prefix='funcappplan', length=24)
+        self.cmd('functionapp plan create -g {} -n {} --sku EP1 --min-instances 4 --max-burst 12' .format(resource_group, plan), checks=[
+            JMESPathCheck('maximumElasticWorkerCount', 12),
+            JMESPathCheck('sku.name', 'EP1'),
+            JMESPathCheck('sku.capacity', 4)
+        ])
+        self.cmd('functionapp plan update -g {} -n {} --min-instances 5 --max-burst 11' .format(resource_group, plan), checks=[
+            JMESPathCheck('maximumElasticWorkerCount', 11),
+            JMESPathCheck('sku.name', 'EP1'),
+            JMESPathCheck('sku.capacity', 5)
+        ])
+        self.cmd('functionapp plan show -g {} -n {} '.format(resource_group, plan), checks=[
+            JMESPathCheck('maximumElasticWorkerCount', 11),
+            JMESPathCheck('sku.name', 'EP1'),
+            JMESPathCheck('sku.capacity', 5)
+        ])
+        self.cmd('functionapp delete -g {} -n {}'.format(resource_group, plan))
+
+
+class FunctionAppServicePlanLinux(ScenarioTest):
+    @ResourceGroupPreparer(location='westus')
+    def test_functionapp_app_service_plan_linux(self, resource_group):
+        plan = self.create_random_name(prefix='funcappplan', length=24)
+        self.cmd('functionapp plan create -g {} -n {} --sku S1 --is-linux' .format(resource_group, plan), checks=[
+            JMESPathCheck('sku.name', 'S1'),
+            JMESPathCheck('kind', 'linux')
+        ])
+
 
 class WebappAuthenticationTest(ScenarioTest):
     @ResourceGroupPreparer(name_prefix='cli_test_webapp_authentication')
@@ -1163,6 +1249,7 @@ class WebappZipDeployScenarioTest(ScenarioTest):
 
 
 class WebappImplictIdentityTest(ScenarioTest):
+    @AllowLargeResponse(8192)
     @ResourceGroupPreparer()
     def test_webapp_assign_system_identity(self, resource_group):
         scope = '/subscriptions/{}/resourcegroups/{}'.format(self.get_subscription_id(), resource_group)
@@ -1176,11 +1263,11 @@ class WebappImplictIdentityTest(ScenarioTest):
             self.cmd('webapp identity show -g {} -n {}'.format(resource_group, webapp_name), checks=[
                 self.check('principalId', result['principalId'])
             ])
+
         self.cmd('role assignment list -g {} --assignee {}'.format(resource_group, result['principalId']), checks=[
             JMESPathCheck('length([])', 1),
             JMESPathCheck('[0].roleDefinitionName', role)
         ])
-
         self.cmd('webapp identity show -g {} -n {}'.format(resource_group, webapp_name), checks=self.check('principalId', result['principalId']))
         self.cmd('webapp identity remove -g {} -n {}'.format(resource_group, webapp_name))
         self.cmd('webapp identity show -g {} -n {}'.format(resource_group, webapp_name), checks=self.is_empty())
@@ -1195,7 +1282,6 @@ class WebappListLocationsFreeSKUTest(ScenarioTest):
 
 
 class WebappTriggeredWebJobListTest(ScenarioTest):
-
     @record_only()
     @ResourceGroupPreparer()
     def test_webapp_triggeredWebjob_list(self, resource_group):