+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_params.py
@@ -1052,7 +1052,7 @@ def load_arguments(self, _):
 
         c.argument('tier',
                    arg_type=tier_param_type,
-                   help='The edition component of the sku. Allowed value is GeneralPurpose.')
+                   help='The edition component of the sku. Allowed values: GeneralPurpose, BusinessCritical.')
 
         c.argument('family',
                    arg_type=family_param_type,
@@ -1076,6 +1076,18 @@ def load_arguments(self, _):
         c.argument('collation',
                    help='The collation of the managed instance.')
 
+        c.argument('proxy_override',
+                   arg_type=get_enum_type(ServerConnectionType),
+                   help='The connection type used for connecting to the instance.')
+
+        c.argument('public_data_endpoint_enabled',
+                   arg_type=get_three_state_flag(),
+                   help='Whether or not the public data endpoint is enabled for the instance.')
+
+        c.argument('timezone_id',
+                   help='The time zone id for the instance to set. '
+                   'A list of time zone ids is exposed through the sys.time_zone_info (Transact-SQL) view.')
+
     with self.argument_context('sql mi create') as c:
         # Create args that will be used to build up the ManagedInstance object
         create_args_for_complex_type(
@@ -1086,7 +1098,10 @@ def load_arguments(self, _):
                 'virtual_network_subnet_id',
                 'vcores',
                 'storage_size_in_gb',
-                'collation'
+                'collation',
+                'proxy_override',
+                'public_data_endpoint_enabled',
+                'timezone_id',
             ])
 
         # Create args that will be used to build up the Managed Instance's Sku object