+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_actions.py
@@ -187,5 +187,5 @@ def _get_latest_image_version(cli_ctx, location, publisher, offer, sku):
                                                                            top=1,
                                                                            orderby='name desc')
     if not top_one:
-        raise CLIError("Can't resolve the vesion of '{}:{}:{}'".format(publisher, offer, sku))
+        raise CLIError("Can't resolve the version of '{}:{}:{}'".format(publisher, offer, sku))
     return top_one[0].name