+++ src/azure-cli/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+2.0.62
+++++++
+* Minor fixes.
+
+2.0.61
+++++++
+* Minor fixes.
 
 2.0.60
 ++++++