+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,7 +2,14 @@
 
 Release History
 ===============
+
+2.5.0
++++++
+* create-for-rbac: handle --years correctly
 * [BREAKING CHANGE] `role assignment delete`: prompt when delete all assignments under the subscription unconditionally
+
+2.4.3
++++++
 * `ad app/sp credential reset`: support --end-date
 * `ad app permission add`: support to add permissions on existing api
 * `ad app permission list`: fix a bad error when there is no permissions