+++ src/command_modules/azure-cli-servicefabric/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.1.16
+++++++
+* Minor fixes.
+
+0.1.15
+++++++
+* Minor fixes.
+
 0.1.14
 ++++++
 * sf cluster list: Fix issue 'ClusterListResult is not iterable.